requirements.txt
tensorflow==2.17.0
num2words
numpy
pandas
transformers
azure-ai-textanalytics
flask
groq